# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/staff:
    get:
      operationId: listStaff
      tags: [staff]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: status, in: query, schema: { type: string } }
      responses:
        "200":
          description: Staff accounts
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Staff" }
    post:
      operationId: createStaff
      tags: [staff]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffInput"
      responses:
        "201":
          description: Created staff account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffIssueResponse"
  /v1/staff/developments:
    get:
      operationId: listStaffDevelopments
      tags: [staff]
      security: [{ bearerAuth: [] }]
      summary: List active developments assignable to staff by the current actor
      responses:
        "200":
          description: Sorted, deduplicated development names
          content:
            application/json:
              schema:
                type: array
                items: { type: string }
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/staff/{id}/reset-code:
    post:
      operationId: resetStaffCode
      tags: [staff]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffCodeUpdate"
      responses:
        "200":
          description: New one-time access code
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StaffIssueResponse" }
        "403":
          description: Not authorized to manage this account
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/staff/{id}/approve:
    post:
      operationId: approveStaff
      tags: [staff]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Approved staff account and one-time access code
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StaffIssueResponse" }
        "403":
          description: Only Human Resources may approve a pending employee
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/staff/{id}:
    delete:
      operationId: deleteStaff
      tags: [staff]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "204":
          description: Staff account permanently deleted
        "403":
          description: Not authorized to delete this account
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/staff/{id}/revoke:
    post:
      operationId: revokeStaff
      tags: [staff]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Revoked staff account
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Staff" }
        "403":
          description: Not authorized to manage this account
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Staff:
      type: object
      required: [id, tenantId, name, role, position, status, developments]
      properties:
        id: { type: string }
        tenantId: { type: string }
        name: { type: string }
        firstName: { type: string, nullable: true }
        lastName: { type: string, nullable: true }
        role:
          type: string
          enum: [administrator, human_resources, management, worker, inspector, procurement, vendor, resident, emergency]
        position:
          type: string
          enum: [Borough Director, Regional Director, Assistant Regional Director, Property Manager, Superintendent, Assistant Superintendent, Supervisor Inspector, Plumbing Supervisor, Electrical Supervisor, Maintenance Supervisor, CPM Supervisor, Grounds Supervisor, Inspector, Plumber, Electrician, Maintenance Worker, Caretaker, Porter, Laborer, Groundskeeper, Administrative Staff, Other Support Staff, Human Resources, Assistant Property Manager, Housing Assistant, Janitorial Staff, CPM, Elevator Service, Painter, Plumber Supervisor, Electric Supervisor, Elevator Supervisor, Painter Supervisor, Carpenter Supervisor, Carpenter, Roofer, General Construction, CCTV Installation, Heating Service, Staff Worker, Director, Other]
        status: { type: string }
        developments:
          type: array
          items: { type: string }
        canManage: { type: boolean }
        canResetCode: { type: boolean }
        canRevoke: { type: boolean }
        canDelete: { type: boolean }
        canApprove: { type: boolean }
...
    StaffIssueResponse:
      allOf:
        - $ref: "#/components/schemas/Staff"
        - type: object
          required: [code]
          properties:
            code: { type: string, minLength: 4, maxLength: 4 }
    StaffInput:
      type: object
      additionalProperties: false
      required: [name, role, position]
      properties:
        name: { type: string }
        firstName: { type: string }
        lastName: { type: string }
        role:
          type: string
          enum: [administrator, human_resources, management, worker, inspector, procurement, vendor, emergency]
        position:
          type: string
          enum: [Borough Director, Regional Director, Assistant Regional Director, Property Manager, Superintendent, Assistant Superintendent, Supervisor Inspector, Plumbing Supervisor, Electrical Supervisor, Maintenance Supervisor, CPM Supervisor, Grounds Supervisor, Inspector, Plumber, Electrician, Maintenance Worker, Caretaker, Porter, Laborer, Groundskeeper, Administrative Staff, Other Support Staff, Human Resources, Assistant Property Manager, Housing Assistant, Janitorial Staff, CPM, Elevator Service, Painter, Plumber Supervisor, Electric Supervisor, Elevator Supervisor, Painter Supervisor, Carpenter Supervisor, Carpenter, Roofer, General Construction, CCTV Installation, Heating Service, Staff Worker, Director, Other]
        developments:
          type: array
          items: { type: string }
        clientRequestId:
          type: string
          minLength: 8
          maxLength: 100
          pattern: "^[a-zA-Z0-9_-]+$"
        status:
          type: string
          enum: [pending, approved]
    StaffCodeUpdate:
      type: object
      additionalProperties: false
      properties: {}
...
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }